+++ ReversingLabs/SDK/helper.py
@@ -78,6 +78,11 @@ class InternalServerError(Exception):
         super(InternalServerError, self).__init__(message)
 
 
+class BadGatewayError(Exception):
+    def __init__(self, message="The server received an invalid response from another server"):
+        super(BadGatewayError, self).__init__(message)
+
+
 class ServiceUnavailableError(Exception):
     def __init__(self, message="Service unavailable"):
         super(ServiceUnavailableError, self).__init__(message)
@@ -89,7 +94,7 @@ class NotAllowedError(Exception):
 
 
 class TooManyRequestsError(Exception):
-    def __init__(self, message="Too many requests - limit reached"):
+    def __init__(self, message="Too many requests. Your quota limit might be reached"):
         super(TooManyRequestsError, self).__init__(message)
 
 
@@ -110,6 +115,7 @@ RESPONSE_CODE_ERROR_MAP = {
     HTTPStatus.CONFLICT: ConflictError,
     HTTPStatus.REQUEST_ENTITY_TOO_LARGE: RequestTooLargeError,
     HTTPStatus.INTERNAL_SERVER_ERROR: InternalServerError,
+    HTTPStatus.BAD_GATEWAY: BadGatewayError,
     HTTPStatus.SERVICE_UNAVAILABLE: ServiceUnavailableError
 }
 
@@ -121,6 +127,9 @@ def validate_hashes(hash_input, allowed_hash_types):
         :param allowed_hash_types: allowed hash types
         :type allowed_hash_types: tuple
     """
+    if not hash_input:
+        raise WrongInputError("At least one hash needs to be supplied as input.")
+
     for hash_string in hash_input:
         try:
             codecs.decode(hash_string, "hex")